RestFul 하게 수정해봤습니다

+++ src/main/java/com/groupware/mapper/NoticeMapper.java
@@ -8,13 +8,13 @@ import com.groupware.vo.SearchVO;
 public interface NoticeMapper {
 	public List<NoticeDTO> selectList(SearchVO searchVO);
 	
-	public int insert(NoticeDTO Notice);
+	public Long insert(NoticeDTO Notice);
 	
 	public NoticeDTO select(Long noticeIdx);
 
-	public int update(NoticeDTO NoticeDTO);
+	public Long update(NoticeDTO NoticeDTO);
 
-	public int delete(Long noticeIdx);
+	public Long delete(Long noticeIdx);
 	
 	public int deleteBySchedule();
 }
